scripts/requirements.txt
pygithub>=2.3.0
python-dateutil>=2.9.0
requests>=2.31.0
